\name{SBSA-package}
\alias{SBSA-package}
\alias{SBSA}
\docType{package}
\title{Simplified Bayesian Sensitivity Analysis}
\description{
Simplified Bayesian sensitivity analysis of models with partially
observed confounders.
}
\details{
  The \code{SBSA} package is an implementation of algorithms for
  simplified Bayesian sensitivity analysis described in \cite{Gustafson
    et al (2010)}. It has one entry function,
  \code{\link{fitSBSA}}. For more details refer to the relevant help files.
}
\author{
  Davor Cubranic \email{cubranic@stat.ubc.ca} and
  Paul Gustafson \email{gustaf@stat.ubc.ca}

  Maintainer: Davor Cubranic \email{cubranic@stat.ubc.ca}
}
\references{
Gustafson, P., L. C. McCandless, A. R. Levy, and S. Richardson. (2010)
\sQuote{Simplified Bayesian Sensitivity Analysis for Mismeasured
      and Unobserved Confounders.}
Biometrics, 66(4):1129-1137.
DOI: 10.1111/j.1541-0420.2009.01377.x
}
\keyword{Bayesian inference}
\keyword{Measurement error}
\keyword{Sensitivity analysis}
\keyword{Unobserved confounder}
\seealso{
\code{\link{fitSBSA}}
}
\examples{
## see examples for fitSBSA
}
